Make Warnings as Errors, upgrade Warning level, fix reported issues
Unused variables mostly.

+++ OpenGLProject/OpenGLProject/Application/Application.cpp
@@ -92,7 +92,7 @@ void Application::Pause(const bool inIsPaused)
 	isPaused = inIsPaused;
 	if (isPaused == false)
 	{
-		glfwSetTime(lastFrameElapsedTime);
+		glfwSetTime(static_cast<double>(lastFrameElapsedTime));
 	}
 }
